requirements.txt
# ============================================================================
# Project - Requirements
# Python 3.8+ required
# ============================================================================

# Web Scraping & Browser Automation
# ----------------------------------------------------------------------------
selenium==4.15.2                    # Browser automation for JavaScript-heavy sites
webdriver-manager==4.0.1            # Automatic ChromeDriver installation
beautifulsoup4==4.12.2              # HTML parsing (backup method)
lxml==4.9.3                         # XML/HTML parser for BeautifulSoup
requests==2.31.0                    # HTTP library

# Computer Vision & Image Processing
# ----------------------------------------------------------------------------
opencv-python==4.8.1.78             # Image processing, color analysis, edge detection
Pillow==10.1.0                      # Image loading, saving, format conversion
scikit-image==0.22.0                # Advanced image processing algorithms
numpy==1.24.3                       # Numerical computing (required by CV libraries)

# Machine Learning & Clustering
# ----------------------------------------------------------------------------
scikit-learn==1.3.2                 # K-Means clustering, dimensionality reduction
scipy==1.11.4                       # Scientific computing, entropy calculations

# Deep Learning (Optional - for advanced features)
# ----------------------------------------------------------------------------
torch==2.1.1                        # PyTorch (for CLIP embeddings if needed)
torchvision==0.16.1                 # Pre-trained segmentation models (DeepLabV3+)

# Time Series & Econometrics (CRITICAL FOR RESEARCH)
# ----------------------------------------------------------------------------
statsmodels==0.14.0                 # ADF test, VAR, Granger Causality, ACF/PACF
pandas==2.0.3                       # Time series manipulation, data alignment
arch==6.2.0                         # GARCH models (if analyzing volatility clustering)

# Financial Data
# ----------------------------------------------------------------------------
yfinance==0.2.32                    # Download VIX, stock data from Yahoo Finance
fredapi==0.5.1                      # Federal Reserve Economic Data (CPI, GDP, CCI)
pandas-datareader==0.10.0           # Alternative source for economic data

# Data Visualization (Publication Quality)
# ----------------------------------------------------------------------------
matplotlib==3.8.2                   # Base plotting library
seaborn==0.13.0                     # Statistical visualizations
plotly==5.18.0                      # Interactive plots (optional for notebooks)

# Jupyter & Notebooks
# ----------------------------------------------------------------------------
jupyter==1.0.0                      # Jupyter notebook
ipykernel==6.27.1                   # Jupyter kernel
notebook==7.0.6                     # Jupyter notebook interface

# Data Management
# ----------------------------------------------------------------------------
openpyxl==3.1.2                     # Excel file support
xlrd==2.0.1                         # Reading old Excel formats

# Utilities
# ----------------------------------------------------------------------------
tqdm==4.66.1                        # Progress bars
python-dotenv==1.0.0                # Environment variable management
pytz==2023.3                        # Timezone handling for date alignment

# Testing (Optional but Recommended)
# ----------------------------------------------------------------------------
pytest==7.4.3                       # Testing framework
pytest-cov==4.1.0                   # Code coverage

# Code Quality (Optional)
# ----------------------------------------------------------------------------
black==23.12.0                      # Code formatting
flake8==6.1.0                       # Linting
isort==5.13.2                       # Import sorting

